Update multinary_infer_sym.cc sliceFlattenView

+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/multiary_infer_sym.cc
@@ -1531,7 +1531,7 @@ bool FlashAttnOpInferSymbolicShape(
 
 static std::vector<symbol::DimExpr> sliceFlattenView(
     const std::vector<symbol::DimExpr> &in_shape,
-    int axis,
+    size_t axis,
     symbol::DimExpr offset,
     symbol::DimExpr sliceLength) {
   PADDLE_ENFORCE_LT(
